// P03-pokedex/src/types/PokemonTypeBG.ts
enum PokemonTypeBG {
    BUG = "bg-[url('../public/assets/backgrounds/bug.jpg')] bg-cover bg-bottom",
    DARK = "bg-[url('../public/assets/backgrounds/dark.jpg')] bg-cover bg-bottom",
    DRAGON = "bg-[url('../public/assets/backgrounds/dragon.jpg')] bg-cover bg-bottom",
    ELECTRIC = "bg-[url('../public/assets/backgrounds/electric.jpg')] bg-cover bg-bottom",
    FAIRY = "bg-[url('../public/assets/backgrounds/fairy.jpg')] bg-cover bg-bottom",
    FIGHTING = "bg-[url('../public/assets/backgrounds/fighting.jpg')] bg-cover bg-bottom",
    FIRE = "bg-[url('../public/assets/backgrounds/fire.jpg')] bg-cover bg-bottom",
    FLYING = "bg-[url('../public/assets/backgrounds/flying.jpg')] bg-cover bg-bottom",
    GHOST = "bg-[url('../public/assets/backgrounds/ghost.jpg')] bg-cover bg-bottom",
    GRASS = "bg-[url('../public/assets/backgrounds/grass.jpg')] bg-cover bg-bottom",
    GROUND = "bg-[url('../public/assets/backgrounds/ground.jpg')] bg-cover bg-bottom",
    ICE = "bg-[url('../public/assets/backgrounds/ice.jpg')] bg-cover bg-bottom",
    NORMAL = "bg-[url('../public/assets/backgrounds/normal.jpg')] bg-cover bg-bottom",
    POISON = "bg-[url('../public/assets/backgrounds/poison.jpg')] bg-cover bg-bottom",
    PSYCHIC = "bg-[url('../public/assets/backgrounds/psychic.jpg')] bg-cover bg-bottom",
    ROCK = "bg-[url('../public/assets/backgrounds/rock.jpg')] bg-cover bg-bottom",
    STEEL = "bg-[url('../public/assets/backgrounds/steel.jpg')] bg-cover bg-bottom",
    WATER = "bg-[url('../public/assets/backgrounds/water.jpg')] bg-cover bg-bottom"
}

export default PokemonTypeBG;